move function into content
